services/rerank/requirements.txt
fastapi==0.115.11
uvicorn==0.34.0
gunicorn==23.0.0 
pydantic==2.10.6
rerankers==0.8.0
transformers==4.48.3
python-dotenv==1.0.1
torch==2.6.0
sentencepiece==0.2.0
protobuf==5.29.3
flashrank==0.2.10
bm25s==0.2.8
PyStemmer==2.2.0.3
httpx==0.28.1